+++ packages/mask/src/plugins/Referral/Worker/apis/entitlements.ts
@@ -3,8 +3,8 @@ import { Interface } from '@ethersproject/abi'
 import { keccak256 } from 'web3-utils'
 
 import { expandEvmAddressToBytes32 } from '../../helpers'
-import { type Entitlement, RpcMethod } from '../../types'
-import { getOracle, rpcCall } from './oracle'
+import type { Entitlement } from '../../types'
+import { getOracle, rpcCall, RpcMethod, RpcRoutes } from './oracle'
 import { supportedOracleChainId } from '../../constants'
 
 const Entitlement = 'Entitlement'
@@ -29,7 +29,9 @@ export async function getAccountEntitlements(account: string): Promise<Entitleme
     const host = await getOracle()
     const topics = [eventIdsEntitlement.Entitlement, '', expandEvmAddressToBytes32(account)]
 
-    const res = await rpcCall(host, RpcMethod.oracle_getLogs, [{ topics, chainId: [supportedOracleChainId] }])
+    const res = await rpcCall(`${host}/v1/${RpcRoutes.rpc}`, RpcMethod.oracle_getLogs, [
+        { topics, chainId: [supportedOracleChainId] },
+    ])
 
     return parseEntitlementEvents(res?.result)
 }
