+++ packages/mask/src/plugins/Referral/Worker/apis/oracle.ts
@@ -3,6 +3,29 @@ import type { JsonRpcResponse } from 'web3-core-helpers'
 import { getFarmOraclesDiscovery } from './discovery'
 import { supportedOracleChainId } from '../../constants'
 
+export enum RpcRoutes {
+    rpc = 'rpc',
+    recommendations = 'recommendations/rpc',
+}
+
+export enum RpcMethod {
+    oracle_chainId = 'oracle_chainId',
+    oracle_getDerivedBlockByHash = 'oracle_getDerivedBlockByHash',
+    oracle_getBundleReceipt = 'oracle_getBundleReceipt',
+    oracle_getDerivedBlockByNumber = 'oracle_getDerivedBlockByNumber',
+    oracle_getOperationalAddress = 'oracle_getOperationalAddress',
+    oracle_call = 'oracle_call',
+    oracle_getLogs = 'oracle_getLogs',
+}
+
+export enum RecommendationsRpcMethod {
+    oracle_getOperationalAddress = 'oracle_getOperationalAddress',
+    oracle_getTimePromise = 'oracle_getTimePromise',
+    oracle_sendProofOfRecommendationOrigin = 'oracle_sendProofOfRecommendationOrigin',
+    oracle_sendProofOfRecommendation = 'oracle_sendProofOfRecommendation',
+    oracle_getProofOfRecommendationOrigin = 'oracle_getProofOfRecommendationOrigin',
+}
+
 export async function getOracle(): Promise<string> {
     const {
         discovery: { oracles },
@@ -31,14 +54,12 @@ export async function jsonReq(url: string, opts: any) {
 
 /**
  *
- * @param host url of the Oracle
+ * @param url url
  * @param method get/post/put
  * @param params params
- * @param porRequest true if this is Proof of Recommendations request (requires different route for max uptime)
  * @returns result of request
  */
- export async function rpcCall(host: string, method: string, params: any, porRequest = false) {
-    const url = porRequest ? `${host}/v1/rpc` : `${host}/v1/recommendations/rpc`
+export async function rpcCall(url: string, method: string, params: any) {
     const res = await jsonReq(url, {
         method: 'POST',
         body: JSON.stringify({
@@ -51,4 +72,3 @@ export async function jsonReq(url: string, opts: any) {
 
     return res
 }
-
